perbaikan tampilan sistem dan menambahkan fitur tangga cut off utk penilaian monev

+++ database/migrations/2025_11_24_070001_create_monev_dynamic_tables.php
@@ -29,7 +29,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->uuid('id_evaluasi_mahasiswa');
             $table->uuid('id_kriteria_monev');
-            $table->integer('nilai');
+            $table->decimal('nilai');
             $table->timestamps();
 
             $table->foreign('id_evaluasi_mahasiswa')->references('id')->on('evaluasi_mahasiswa')->onDelete('cascade');
